requirements.txt
biopython==1.81
bioc==2.0.post5
flask==2.2.3
gensim==4.3.0
matplotlib==3.7.1
nltk==3.8.1
numpy==1.23.5
openpyxl==3.1.1
pandas==1.5.3
scikit-learn==1.2.2
seaborn==0.12.2
tensorflow-gpu==2.10.0
tqdm==4.65.0
yellowbrick==1.5
